investigating typecast issue

+++ packages/ceti-tag-data-capture/src/max17320.h
@@ -97,21 +97,21 @@ typedef struct {
 // Device struct
 typedef struct __MAX17320_HandleTypeDef {
 	max17320_Reg_Status status;
-	float remaining_capacity; // mAh
-    float state_of_charge; // %
+	uint16_t remaining_capacity; // mAh
+    uint16_t state_of_charge; // %
 
-    float cell_1_voltage; // Voltage of cell connected between MID and NEG
-    float cell_2_voltage; // Voltage of cell connected between POS and MID
-    float total_battery_voltage; // Total battery pack voltage between POS and NEG
-    float pack_side_voltage; // Pack-side voltage input into system
+    uint16_t cell_1_voltage; // Voltage of cell connected between MID and NEG
+    uint16_t cell_2_voltage; // Voltage of cell connected between POS and MID
+    uint16_t total_battery_voltage; // Total battery pack voltage between POS and NEG
+    uint16_t pack_side_voltage; // Pack-side voltage input into system
 
-    float temperature; // °C
+    int16_t temperature; // °C
 
-    float battery_current; // A
-    float average_current; // A
+    int16_t battery_current; // A
+    int16_t average_current; // A
 
-    float time_to_empty; // h
-    float time_to_full; // h
+    uint16_t time_to_empty; // h
+    uint16_t time_to_full; // h
 
 	uint8_t remaining_writes;
 } MAX17320_HandleTypeDef;
